Minor Improvements
Grabbing a shield powerup now displays a visual around your ship so that you know it is active. Switched several instance variables in PlayerController and EnemyBehavior from public to private to prevent confusion. Tweaked the spawn mechanic for enemies so that they will not spawn continuously if you shoot them before their entry animation finishes.  At present, this is the last change I have planned for this project.